--- src/App.js
@@ -1,34 +1,0 @@
-import React, { Component } from 'react';
-import CardList from './CardList';
-import { robots } from './robots';
-import SearchBox from './SearchBox';
-import './App.css'; 
-
-class App extends Component {
-	constructor() {
-		super()
-		// State is something, that can CHANGE
-		this.state = {
-			robots: robots,
-			searchfield: ''
-		}
-	}
-	onSearchChange = (event) => {
-		//Update state (searchfield in our case). If you want to change state, you have to always do this...
-		this.setState({searchfield: event.target.value})
-	}
-	render() {
-		const filteredRobots = this.state.robots.filter(robots => {
-			return robots.name.toLowerCase().includes(this.state.searchfield.toLowerCase());
-		})
-		return (
-		<div className='tc'>
-			<h1 className='f1'>Robofriends</h1>
-			<SearchBox searchChange={this.onSearchChange}/>
-			<CardList robots={filteredRobots}/>
-		</div>
-	  );
-	}
-}
-
-export default App;
